refactor: polished code, removed warnings
