+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "private-CMTA-Token",
+  "name": "CMTAT",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
