+++ src/test/reading_constants.nr
@@ -1,6 +1,5 @@
 use crate::test::utils;
-use dep::aztec::test::helpers::cheatcodes;
-use crate::Token;
+use crate::CMTAToken as Token;
 
 // It is not possible to deserialize strings in Noir ATM, so name and symbol cannot be checked yet.
