+++ src/test/test_pause_module.nr
@@ -1,13 +1,11 @@
 use crate::test::utils;
-use dep::aztec::{test::helpers::cheatcodes};
-use dep::authwit::cheatcodes as authwit_cheatcodes;
-use crate::Token;
-use crate::types::access_controlModule::{DEFAULT_ADMIN_ROLE, PAUSE_ROLE, MINTER_ROLE};
+use crate::CMTAToken as Token;
+use crate::modules::access_controlModule::{PAUSE_ROLE, MINTER_ROLE};
 
 
 #[test]
 unconstrained fn get_paused(){
-    let (env, token_contract_address, issuer, _) = utils::setup(/* with_account_contracts */ false);
+    let (env, token_contract_address, _, _) = utils::setup(/* with_account_contracts */ false);
 
     let ret = Token::at(token_contract_address).public_get_pause().view(&mut env.public());
     assert(ret == 0);
@@ -15,7 +13,7 @@ unconstrained fn get_paused(){
 
 #[test]
 unconstrained fn pause_test(){
-    let (env, token_contract_address, issuer, user) = utils::setup(/* with_account_contracts */ false);
+    let (env, token_contract_address, issuer, _) = utils::setup(/* with_account_contracts */ false);
 
     //check if issuer has pause role
     let ret = Token::at(token_contract_address).has_role(PAUSE_ROLE, issuer).view(&mut env.public());
@@ -53,7 +51,7 @@ unconstrained fn pause_test(){
 
 #[test(should_fail_with="Error: token contract is paused")]
 unconstrained fn mint_when_paused_fails(){
-    let (env, token_contract_address, issuer, user) = utils::setup(/* with_account_contracts */ false);
+    let (env, token_contract_address, issuer, _) = utils::setup(/* with_account_contracts */ false);
 
     let mint_amount = 10000;
     env.impersonate(issuer);
@@ -72,7 +70,7 @@ unconstrained fn mint_when_paused_fails(){
 
 #[test(should_fail_with="Error: token contract is paused")]
 unconstrained fn transfer_when_paused_fails(){
-    let (env, token_contract_address, issuer, user,_) = utils::setup_and_mint(/* with_account_contracts */ false);
+    let (env, token_contract_address, issuer, _,_) = utils::setup_and_mint(/* with_account_contracts */ false);
 
     let mint_amount = 100;
     env.impersonate(issuer);
@@ -93,7 +91,7 @@ unconstrained fn transfer_when_paused_fails(){
 #[test(should_fail_with="error: contract already paused")]
 unconstrained fn pause_twice_throws(){
 
-    let (env, token_contract_address, issuer, user) = utils::setup(/* with_account_contracts */ false);
+    let (env, token_contract_address, issuer, _) = utils::setup(/* with_account_contracts */ false);
     
     Token::at(token_contract_address).grant_role(PAUSE_ROLE, issuer).call(&mut env.public());
 
@@ -109,7 +107,7 @@ unconstrained fn pause_twice_throws(){
 //pause fails when caller does not have pause role
 #[test(should_fail_with="AccessControlUnauthorizedAccount")]
 unconstrained fn test_pause_roles(){
-    let (env, token_contract_address, issuer, user) = utils::setup(/* with_account_contracts */ false);
+    let (env, token_contract_address, _, _) = utils::setup(/* with_account_contracts */ false);
     
     Token::at(token_contract_address).pause_contract().call(&mut env.public());
 }
