+++ src/types.nr
@@ -1,8 +1,3 @@
-pub mod access_controlModule;
-pub mod validationModule;
-pub mod pauseModule;
-pub mod enforcementModule;
-pub mod extensions;
 pub mod token_note;
